Add files via upload
We perform this small analysis on a dataset published in Kaggles. 

Context
Retail dataset of a global superstore for 4 years.
Perform EDA and Predict the sales of the next 7 days from the last date of the Training dataset!

Content
Time series analysis deals with time series based data to extract patterns for predictions and other characteristics of the data. It uses a model for forecasting future values in a small time frame based on previous observations. It is widely used for non-stationary data, such as economic data, weather data, stock prices, and retail sales forecasting.

Dataset
The dataset is easy to understand and is self-explanatory

Inspiration:
- Understanding key sales
- Visualize the regions with the highest sales rates
- Understand which products are most in demand